+++ clients/derivatives_trading_coin_futures/src/binance_sdk_derivatives_trading_coin_futures/rest_api/rest_api.py
@@ -2272,9 +2272,9 @@ class DerivativesTradingCoinFuturesRestAPI:
                     new_order_resp_type (Optional[NewOrderNewOrderRespTypeEnum] = None):
                     price_match (Optional[NewOrderPriceMatchEnum] = None): only avaliable for `LIMIT`/`STOP`/`TAKE_PROFIT` order;
         Can't be passed together with `price`
-                    self_trade_prevention_mode (Optional[NewOrderSelfTradePreventionModeEnum] = None): `EXPIRE_TAKER`:expire taker order when STP triggers/
-        `EXPIRE_MAKER`:expire taker order when STP triggers/
-        `EXPIRE_BOTH`:expire both orders when STP triggers
+                    self_trade_prevention_mode (Optional[NewOrderSelfTradePreventionModeEnum] = None): `EXPIRE_TAKER`: expire taker order when STP triggers/
+        `EXPIRE_MAKER`: expire taker order when STP triggers/
+        `EXPIRE_BOTH`: expire both orders when STP triggers
                     recv_window (Optional[int] = None):
 
                 Returns: